code/server/requirements.txt
Django
mysql-connector-python
